+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 2018-10-31
+
+* migrated to python3
+* introduced documentation per method
+* add event(s) to a sketch
+** added user_input to provice sketch id
+** added ability to use the current system time as timestamp
+** added check for right timestamp format
+* Adjusted requirements.txt
+* adjusted .gitignore
+* version 0.1
+
 # 2018-10-30
 
 * initial version